[test] New test mode to execute all tests with C++ interop enabled.
To improve the reliability of the C++ interop, a new test mode is added
to the existing ones. The new mode enables the usage of
`--enable-experimental-cxx-interop` in every test invocation of the frontend and the
driver.

The ideal situation would be that with and without the C++ interop the
test should not fail. Currently around 283 tests seems to still fail.