Disallow async IBAction Methods
The IBAction signature does not allow completion handlers, so making
IBAction methods async doesn't make sense. Disallowing it statically so
it doesn't just break people when they try.

This adds a note with a possible fix + fix-it for the issue,
recommending removing the `async` and wrapping the body of the function
in a task.