[lit] Support executable and non-executable tests with C++ Interop
In #59120 I did not realize that I was filtering tests that required
executable or non-executable features, so many tests were being skipped.

This change should recover both sets of tests and it allows us to test
C++ interop in a lot more cases.

Sadly this raises the number of failing tests with C++ interop to 425.