[capture-promotion] Add support for GEPs from captures that are load_borrowed.
We supported this code pattern previously just for loads and never implemented
the support for load_borrow. I just added analogous code for
load_borrow. Without this, we hit a crash since:

1. We would not process the struct_element_addr for a valid value.
2. The load_borrow would not look through the struct_element_addr, so we would
ask the cloner for the mapped value to the struct_element_addr... it was never
visited, thus never mapped, thus KABOOM.

rdar://93879907