[BackDeployConcurrency] Fix to use the new threading library.
A copy of the older Concurrency sources was just added to BackDeployConcurrency,
replacing the `add_subdirectory(../Concurrency)` trick it was using previously.
These need to be adjusted to work with the threading library.  Fortunately, the
necessary adjustments are the same as those required for the existing
Concurrency library, give or take.

rdar://90776105