[move-only] Make SILType::getRawASTType() private and make SILPrinter a friend so we can invoke that method.
The only place outside of SILType that uses getRawASTType() is in SILPrinter. It
is really an inner detail of SILType that we don't want to expose. With that in
mind, in this commit we:

1. Made SILType::getRawASTType() private.

2. Forward declared SILPrinter in SILType.h. We don't define it so no one can
use it, but the declaration still lets us make it a friend class of SILType
(allowing SILPrinter to still access getRawASTType()).