Update Code of Conduct: Text added, format & index
I would like to thanks Apple and Swift, for providing opensource tools as nice as Swift is, this is my little piece of contribution, I have been investing some of my time, in improving this repo, by formatting the text, and indexing it. It is my hope that this will help to improve people's attitudes for a healthier society, with more values and awareness. Thank you Apple.