[move-function-value] When searching for uses to emit notes about in a sub-loop from the move, skip destroy_value.
Once we have detected that we have a use after a move, we do a second pass over
the function looking for uses that are later than the move but are in
sub-loop. In such a case, the use will not be a boundary use but will be LiveOut
due to the loop circularity. In such a case, we need to match what we do
normally with pruned liveness, namely ignore destroy_value uses.

This fixes a false positive note where we emitted an additional remark on a
destroy_value.