add --all-repositories flag
Sometimes when working with Docker to debug a Linux issue, we want
to copy the sources directory into the Docker image. If we have a
checkout that was done on macOS, then repos needed to build those
same sources on Linux, such as `icu` will not be included. There
was no way to ask `update-checkout` to download those repos while
on macOS.

With `--all-repositories`, you can now ask for those repos, despite
them not being needed on macOS. You can even do

`--all-repositories --skip-repository icu` to include all platform
specific repositories, but skip `icu`.