Implement heuristic to prioritize in field descriptor search
Currently when looking for field descriptors we parse the reflection
metadata in whatever order it was registered. This patch implements a
heuristic where we try to match a new optional Name field with the
module name of the type's field descriptor we're looking for.

rdar://87889973
(cherry picked from commit 632d18795ff590e039a72b99be05f0547c8a663e)