Implement heuristic to prioritize in field descriptor search
Currently when looking for field descriptors we parse the reflection
metadata in whatever order it was registered. This patch implements a
heuristic where we try to match a new optional Name field with the
module name of the type's field descriptor we're looking for.

rdar://87889973
(cherry picked from commit 65a9e7de1e49b27ae160211c64434b958619f520)