[Distributed] Fix `SILDeclRef` of distributed thunk to reference originator
Setting `asDistributed()` on a reference to a distributed member
would case `async` and `throws` added to SILFunctionType but the
mangling of the thunk would be based on originator which allows
users to overload distributed methods on effects.

The (big) side-effect of this design  - adding/removing `async`
or `throws` to/from the existing distributed method is no longer
transparent for remote users because it would change the mangling
of the thunk.